+++ src/main/java/com/cova/quizapp/unittest/controller/TriviaController.java
@@ -1,8 +1,7 @@
 package com.cova.quizapp.unittest.controller;
 
 import com.cova.quizapp.service.ITriviaService;
-import com.cova.quizapp.serviceimpl.TriviaServiceImpl;
-import lombok.Data;
+import com.cova.quizapp.unittest.serviceimpl.TriviaServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
