+++ src/main/java/com/cova/quizapp/unittest/controller/TriviaHistoryController.java
@@ -1,7 +1,7 @@
 package com.cova.quizapp.unittest.controller;
 
 import com.cova.quizapp.service.ITriviaHistoryService;
-import com.cova.quizapp.serviceimpl.TriviaHistoryServiceImpl;
+import com.cova.quizapp.unittest.serviceimpl.TriviaHistoryServiceImpl;
 import lombok.Data;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
