+++ src/test/java/com/cova/quizapp/QuizApplicationTest.java
@@ -1,9 +1,15 @@
 package com.cova.quizapp;
 
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
 @SpringBootTest
 public class QuizApplicationTest {
+
+    @Test
+    public void context(){
+
+    }
 }
