+++ src/test/java/com/cova/quizapp/unittest/serviceimpl/TriviaServiceImplTest.java
@@ -1,10 +1,8 @@
-package com.cova.quizapp.serviceimpl;
+package com.cova.quizapp.unittest.serviceimpl;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 class TriviaServiceImplTest {
